Update register Admin and Validate form add and update, Alert delete.

+++ src/main/java/com/example/tnguyenpnv23a_finalproject/Main.java
@@ -16,6 +16,9 @@ import javafx.scene.layout.HBox;
 import javafx.scene.layout.VBox;
 import javafx.stage.Stage;
 import java.util.ArrayList;
+import java.util.Optional;
+import java.util.regex.Matcher;
+import java.util.regex.Pattern;
 
 import javafx.scene.image.Image;
 import javafx.scene.image.ImageView;
@@ -23,9 +26,9 @@ import javafx.scene.image.ImageView;
 public class Main extends Application {
     private Scene homepage, loginScreen, registerScreen, addScreen, updateScreen;
     private Stage window;
-    private VBox formUpdate, formAdd, bookVBox ;
+    private VBox formUpdate, formAdd, bookVBox, formRegister, formLogin ;
     private HBox root;
-    DBConnection DB = new DBConnection();
+    DBConnection DB = new DBConnection(); //tao doi tuong Db connect moi
     private static final String EMPTY = "";
     public static void main(String[] args) {
         launch(args);
@@ -36,14 +39,18 @@ public class Main extends Application {
 
         root = new HBox(10);
 
-        VBox formLogin = new VBox();
         formAdd = new VBox();
+        formRegister = new VBox();
+        formLogin = new VBox();
         VBox controlScreen = new VBox(10);
         VBox bookHBox = new VBox(10);
         bookVBox = new VBox();
         formUpdate = new VBox();
 
+
+
         Button btnLogOut = new Button("Log Out");
+        btnLogOut.setStyle("-fx-background-color: #FF6600");
         btnLogOut.setPadding(new Insets(5, 15, 5, 15));
         btnLogOut.setOnAction(new EventHandler<ActionEvent>() {
             @Override
@@ -52,6 +59,7 @@ public class Main extends Application {
             }
         });
         Button btnAdd = new Button("Add");
+        btnAdd.setStyle("-fx-background-color: #6666FF");
         btnAdd.setPadding(new Insets(5, 15, 5, 15));
         btnAdd.setOnAction(new EventHandler<ActionEvent>() {
             @Override
@@ -68,15 +76,19 @@ public class Main extends Application {
         showLoginPage(formLogin);
         displayHomePage(bookVBox);
         showFormAdd(formAdd);
+        showRegisterForm(formRegister);
 
 
-        loginScreen = new Scene(formLogin,1200, 500);
-        homepage = new Scene(root,1500, 1600);
-        addScreen = new Scene(formAdd,1500, 1600);
-        updateScreen = new Scene(formUpdate,1500, 1600);
+        loginScreen = new Scene(formLogin,1200, 800);
+        homepage = new Scene(root,1200, 800);
+        addScreen = new Scene(formAdd,1200, 800);
+        updateScreen = new Scene(formUpdate,1200, 800);
+        registerScreen = new Scene(formRegister,1200, 800);
+
+//        loginScreen.setFill(Color.web("#C6E2FF"));
 
         window.setTitle("Book");
-        window.setScene(loginScreen);
+        window.setScene(homepage);
         window.show();
     }
     public void displayHomePage (VBox vBox ) {
@@ -106,8 +118,8 @@ public class Main extends Application {
             Image image = new Image(bookList.get(i).getImage());
             ImageView imageView = new ImageView();
             imageView.setImage(image);
-            imageView.setFitWidth(110);
-            imageView.setFitHeight(110);
+            imageView.setFitWidth(130);
+            imageView.setFitHeight(140);
 
             grid.add(new Label("" + (i + 1)), 0, i + 2);
             grid.add(new Label(bookList.get(i).getName()), 1, i + 2);
@@ -119,6 +131,7 @@ public class Main extends Application {
             grid.add(new Label(bookList.get(i).getDescription()), 7, i + 2);
             int finalI = i;
             Button btnEdit = new Button("Edit");
+            btnEdit.setStyle("-fx-background-color: #FFFF33");
             btnEdit.setPadding(new Insets(5, 15, 5, 15));
             btnEdit.setOnAction(e -> {
                 formUpdate(formUpdate, bookList.get(finalI));
@@ -126,20 +139,45 @@ public class Main extends Application {
             });
             grid.add(btnEdit, 8, i + 2);
             Button btnDelete = new Button("Delete");
+            btnDelete.setStyle("-fx-background-color: #FF3300");
             btnDelete.setPadding(new Insets(5, 15, 5, 15));
             btnDelete.setOnAction(new EventHandler<ActionEvent>() {
                 @Override
                 public void handle(ActionEvent event) {
-                    DB.deleteBook(bookList.get(finalI).getId());
-                    vBox.getChildren().clear();
-                    displayHomePage(vBox);
+                    Alert alert = new Alert(Alert.AlertType.CONFIRMATION);
+                    alert.setTitle("Delete File");
+                    alert.setHeaderText("Are you sure want to move this item to the Recycle Bin?");
+                    alert.setContentText("C:/MyFile.txt");
+                    alert.showAndWait().ifPresent(response -> {
+                        if (response == ButtonType.OK) {
+                            Alert alert2 = new Alert(Alert.AlertType.INFORMATION);
+                            alert2.setTitle("Inform");
+                            alert2.setHeaderText("This item is moved to the Recycle Bin!");
+                            alert2.setContentText("Continue!");
+                            DB.deleteBook(bookList.get(finalI).getId());
+                            vBox.getChildren().clear();
+                            displayHomePage(vBox);
+                            alert2.show();
+                        } else {
+                            Alert alert2 = new Alert(Alert.AlertType.INFORMATION);
+                            alert2.setTitle("Inform");
+                            alert2.setHeaderText("This item isn't moved to the Recycle Bin!");
+                            alert2.setContentText("Continue!");
+                            alert2.show();
+                        }
+                    });
                 }
             });
             grid.add(btnDelete, 9, i + 2);
         }
-        vBox.getChildren().add(grid);
+        ScrollPane scrollHomepage = new ScrollPane(grid);
+        scrollHomepage.setHbarPolicy(ScrollPane.ScrollBarPolicy.ALWAYS);
+        scrollHomepage.setVbarPolicy(ScrollPane.ScrollBarPolicy.NEVER);
+        scrollHomepage.setContent(grid);
+        vBox.getChildren().add(scrollHomepage);
     }
     public void showLoginPage (VBox vbox) {
+        vbox.getChildren().clear();
         GridPane grid = new GridPane();
         grid.setAlignment(Pos.CENTER);
         grid.setVgap(10);
@@ -159,17 +197,18 @@ public class Main extends Application {
         fieldPass.setSpacing(10);
         fieldPass.setAlignment(Pos.BASELINE_CENTER);
         Button btnGoBack = new Button("Register");
+        btnGoBack.setStyle("-fx-background-color: #FF9966");
         btnGoBack.setOnAction(actionEvent -> {
             window.setScene(registerScreen);
         });
         Button btnLogin = new Button("LOGIN");
+        btnLogin.setStyle("-fx-background-color: #008800");
         btnLogin.setOnAction(actionEvent -> {
            if (this.checkLogin(name, pass)) {
                Alert alert = new Alert(Alert.AlertType.INFORMATION);
                alert.setTitle("Login");
-               alert.setHeaderText("Hi "+name.getText());
+               alert.setHeaderText("Hello "+name.getText());
                alert.setContentText("Login successfully!");
-//               alert.show();
                alert.showAndWait().ifPresent(response -> {
                    if (response == ButtonType.OK) {
                        name.setText("");
@@ -208,6 +247,8 @@ public class Main extends Application {
         grid.add(tfImage, 1, 1);
         //
         ChoiceBox choiceCategory = new ChoiceBox();
+
+        //lay category trong mang category cho hien ra cac option
         ArrayList<Category> listCategories = DB.getListCategory();
         for (var category : listCategories){
             choiceCategory.getItems().add(category.getName());
@@ -240,26 +281,40 @@ public class Main extends Application {
                 window.setScene(homepage);
             }
         });
-        Button btnSave = new Button("save");
+        Button btnSave = new Button("Save");
         btnSave.setPadding(new Insets(5, 15, 5, 15));
         btnSave.setOnAction(new EventHandler<ActionEvent>() {
             @Override
             public void handle(ActionEvent event) {
-                String name = tfName.getText();
-                String image = tfImage.getText();
-                int idCategory = choiceCategory.getSelectionModel().getSelectedIndex()+1;
-                String author = tfAuthor.getText();
-                Integer price = Integer.valueOf(tfPrice.getText());
-                Integer quantity = Integer.valueOf(tfQuantity.getText());
-                String description = tfDescription.getText();
-                if (!name.equals(EMPTY) && !image.equals(EMPTY) && !price.equals(EMPTY) && !description.equals(EMPTY)) {
-                    DB.insertBook(new Book(name, image, new Category(idCategory), author, price, quantity, description));
-                    displayHomePage(bookVBox);
-                    window.setScene(homepage);
+                if (checkEmpty(tfName) && checkEmpty(tfImage) && checkEmpty(tfPrice) && checkEmpty(tfDescription) && checkEmpty(tfQuantity)) {
+                    if (checkFormat(tfPrice.getText())){
+                        if (checkFormat(tfQuantity.getText())){
+                            String name = tfName.getText();
+                            String image = tfImage.getText();
+                            int idCategory = choiceCategory.getSelectionModel().getSelectedIndex()+1;
+                            String author = tfAuthor.getText();
+                            Integer price = Integer.valueOf(tfPrice.getText());
+                            Integer quantity = Integer.valueOf(tfQuantity.getText());
+                            String description = tfDescription.getText();
+                            DB.insertBook(new Book(name, image, new Category(idCategory), author, price, quantity, description));
+                            displayHomePage(bookVBox);
+                            window.setScene(homepage);
+                        }else {
+                            var alert = new Alert(Alert.AlertType.INFORMATION);
+                            alert.setHeaderText(null);
+                            alert.setContentText("Please input correct type of Quantity!");
+                            alert.showAndWait();
+                        }
+                    }else {
+                        var alert = new Alert(Alert.AlertType.INFORMATION);
+                        alert.setHeaderText(null);
+                        alert.setContentText("Please input correct type of Price!");
+                        alert.showAndWait();
+                    }
                 }else {
                     var alert = new Alert(Alert.AlertType.INFORMATION);
                     alert.setHeaderText(null);
-                    alert.setContentText("Please fill all blank!");
+                    alert.setContentText("Please fill in all blank!");
                     alert.showAndWait();
                 }
             }
@@ -330,21 +385,35 @@ public class Main extends Application {
         btnUpdate.setOnAction(new EventHandler<ActionEvent>() {
             @Override
             public void handle(ActionEvent event) {
-                String name = tfName.getText();
-                String image = tfImage.getText();
-                int idCategory = choiceCategory.getSelectionModel().getSelectedIndex()+1;
-                String author = tfAuthor.getText();
-                Integer price = Integer.valueOf(tfPrice.getText());
-                Integer quantity = Integer.valueOf(tfQuantity.getText());
-                String description = tfDescription.getText();
-                if (!name.equals(EMPTY) && !image.equals(EMPTY) && !price.equals(EMPTY) && !description.equals(EMPTY)) {
-                    DB.updateBook(book.getId(), new Book(name, image, new Category(idCategory), author, price, quantity, description));
-                    displayHomePage(bookVBox);
-                    window.setScene(homepage);
+                if (checkEmpty(tfName) && checkEmpty(tfImage) && checkEmpty(tfPrice) && checkEmpty(tfDescription) && checkEmpty(tfQuantity)) {
+                    if (checkFormat(tfPrice.getText())){
+                        if (checkFormat(tfQuantity.getText())){
+                            String name = tfName.getText();
+                            String image = tfImage.getText();
+                            int idCategory = choiceCategory.getSelectionModel().getSelectedIndex()+1;
+                            String author = tfAuthor.getText();
+                            Integer price = Integer.valueOf(tfPrice.getText());
+                            Integer quantity = Integer.valueOf(tfQuantity.getText());
+                            String description = tfDescription.getText();
+                            DB.updateBook(book.getId(), new Book(name, image, new Category(idCategory), author, price, quantity, description));
+                            displayHomePage(bookVBox);
+                            window.setScene(homepage);
+                        }else {
+                            var alert = new Alert(Alert.AlertType.INFORMATION);
+                            alert.setHeaderText(null);
+                            alert.setContentText("Please input correct type of Quantity!");
+                            alert.showAndWait();
+                        }
+                    }else {
+                        var alert = new Alert(Alert.AlertType.INFORMATION);
+                        alert.setHeaderText(null);
+                        alert.setContentText("Please input correct type of Price!");
+                        alert.showAndWait();
+                    }
                 }else {
                     var alert = new Alert(Alert.AlertType.INFORMATION);
                     alert.setHeaderText(null);
-                    alert.setContentText("Please fill all blank!");
+                    alert.setContentText("Please fill in all blank!");
                     alert.showAndWait();
                 }
             }
@@ -354,17 +423,110 @@ public class Main extends Application {
         vbox.getChildren().add(grid);
     }
 
-    public boolean checkLogin(TextField adminName, TextField password) {
-        ArrayList<Admin> ad = new ArrayList<Admin>();
-        ad = (ArrayList<Admin>) DB.getAdmin();
-        String inputName = adminName.getText();
-        String inputPass = password.getText();
-        if(inputName.equals(ad.get(0).name)&& inputPass.equals(ad.get(0).password)){
-            return true;
-        }else{
-            return false;
-        }
+    public void showRegisterForm (VBox registerView) {
+        GridPane grid = new GridPane();
+        grid.setAlignment(Pos.CENTER);
+        grid.setVgap(10);
+        grid.setHgap(10);
+        Label labelRegister =new Label("REGISTER");
+        Label LabelName = new Label("Name: ");
+        Label LabelPassword = new Label("Password: ");
+        Label LabelRePassword = new Label("RePassword: ");
+        TextField name = new TextField();
+        PasswordField password= new PasswordField();
+        PasswordField rePassword= new PasswordField();
+        Button btnRegister = new Button("Register");
+        Button btnLogin = new Button("Back Login");
+        btnRegister.setOnAction(new EventHandler<ActionEvent>() {
+            @Override
+            public void handle(ActionEvent event) {
+                if (checkEmpty(name) && checkEmpty(password) && checkEmpty(rePassword)){
+                    if (checkCorrectRePass(password.getText(), rePassword.getText())){
+                        DB.registerAdmin(new Admin(name.getText(), password.getText()));
+                        Alert alert = new Alert(Alert.AlertType.INFORMATION);
+                        alert.setTitle("Register");
+                        alert.setHeaderText("Congratulations! Register successfully!");
+                        alert.setContentText("Please Login");
+                        alert.showAndWait().ifPresent(response -> {
+                            if (response == ButtonType.OK) {
+                                window.setScene(loginScreen);
+                            }
+                        });
+                    }else {
+                        Alert alert = new Alert(Alert.AlertType.INFORMATION);
+                        alert.setTitle("Error");
+                        alert.setHeaderText("Password is not correct with RePassword");
+                        alert.setContentText("Please enter password again.");
+                        alert.show();
+                    }
+                }else {
+                    Alert alert = new Alert(Alert.AlertType.INFORMATION);
+                    alert.setTitle("Error");
+                    alert.setHeaderText("Empty Input");
+                    alert.setContentText("Please enter all field.");
+                    alert.show();
+                }
+
+            }
+        });
+
+        btnLogin.setOnAction(new EventHandler<ActionEvent>() {
+            @Override
+            public void handle(ActionEvent event) {
+                name.setText("");
+                password.setText("");
+                rePassword.setText("");
+                window.setScene(loginScreen);
+            }
+        });
+        grid.add(labelRegister, 0,0);
+        grid.add(LabelName, 0, 1);
+        grid.add(name, 1,1);
+        grid.add(LabelPassword, 0, 2);
+        grid.add(password, 1, 2);
+        grid.add(LabelRePassword, 0, 3);
+        grid.add(rePassword, 1, 3);
+        grid.add(btnLogin, 0, 4);
+        grid.add(btnRegister, 1,4);
+        registerView.getChildren().add(grid);
+        registerView.setSpacing(15);
+        registerView.setAlignment(Pos.BASELINE_CENTER);
     }
 
+    public boolean checkEmpty(TextField tf){
+        if (tf.getText().isEmpty()){
+            return false;
+        }
+        return true;
+    }
+    public boolean checkCorrectRePass(String pass, String rePass){
+        if(pass.equals(rePass)){
+            return true;
+        }
+        return false;
+    }
+    public boolean checkLogin(TextField adminName, TextField password) {
+        ArrayList<Admin> ad = DB.getAdmin();
+        for(var admin : ad){
+            System.out.printf(admin.name);
+            System.out.printf(admin.password);
+            if(adminName.getText().equals(admin.name) && password.getText().equals(admin.password)){
+                return true;
+            }
+        }
+        return false;
+    }
+    public boolean checkFormat (String  value) {
+        String regex = "[-+]?(\\d*\\.\\d+|\\d+)";
+        Pattern p = Pattern.compile(regex);
 
+        // Creates a matcher that will match input1 against regex
+        Matcher m = p.matcher(value);
+
+        // If match found and equal to input1
+        if(m.find() && m.group().equals(value)){
+            return true;
+        }
+        return false;
+    }
 }
